Add Recorder

+++ gymutils/observation/concat.py
@@ -8,7 +8,7 @@ def concat_horizontally(observations, margin_width=0):
     for observation in observations[1:]:
         result, observation = pad_to_fit_the_height(result, observation)
         height, _, _ = result.shape
-        margin = np.zeros([height, margin_width, 3])
+        margin = np.zeros([height, margin_width, 3], dtype=result.dtype)
         result = np.concatenate([result, margin, observation], axis=1)
     return result
 
@@ -18,6 +18,6 @@ def concat_vertically(observations, margin_height=0):
     for observation in observations[1:]:
         result, observation = pad_to_fit_the_width(result, observation)
         _, width, _ = result.shape
-        margin = np.zeros([margin_height, width, 3])
+        margin = np.zeros([margin_height, width, 3], dtype=result.dtype)
         result = np.concatenate([result, margin, observation], axis=0)
     return result
